einj_mem_uc: Fix vtop failed in "instr" test case
An error occurs during the execution of the einj_mem_uc "instr" test
and exits. The error message shows that the instr memory page is not
found, and the einj_mem_uc exits. The detailed log is as follows:

$ ./einj_mem_uc -c 10 -f instr
page not present
0: instr    vaddr = 0x403000 paddr = ffffffffffffffff
~/einj_mem_uc : write error on '/sys/kernel/debug/apei/einj/error_inject'

Test Environment:
OS Version:
CentOS Linux release 7.9.2009 (Core)
Kernel Version:
3.10.0-1160.el7.x86_64 #1 SMP Mon Oct 19 16:18:59 UTC 2020 x86_64

Solution: Call the dosum function in instr_alloc to load the dosum
memory page in advance to prevent vtop conversion failure.

Signed-off-by: Zhongyu Gao <gzy@sangfor.com.cn>
Signed-off-by: Tony Luck <tony.luck@intel.com>