einj_mem_uc: Wait longer for patrol scrub CMCI
It may take up to 20 seconds for the patrol scrubber to restart
and scan to the specific location where the error was injected.

Add a flag to indicate that the test should wait for much longer
to check for patrol scrub CMCI.

Also change the message to print the actual delay in units of
seconds, not microseconds (since the values are large enough
that this is a more human readable format).

Signed-off-by: Tony Luck <tony.luck@intel.com>